ISSUES-4006 split mysql protocol
