+++ src/Core/MySQL/MySQLPackets.h
@@ -2,6 +2,9 @@
 
 #include <Core/MySQL/IMySQLReadPacket.h>
 #include <Core/MySQL/IMySQLWritePacket.h>
+#include <Core/MySQL/PacketsConnection.h>
+#include <Core/MySQL/PacketsProtocolText.h>
+#include <Core/MySQL/Authentication.h>
 #include <IO/WriteHelpers.h>
 
 namespace DB
@@ -19,95 +22,6 @@ public:
 
 protected:
     void readPayloadImpl(ReadBuffer & buf) override;
-};
-
-class Handshake : public IMySQLWritePacket, public IMySQLReadPacket
-{
-public:
-    int protocol_version = 0xa;
-    String server_version;
-    uint32_t connection_id;
-    uint32_t capability_flags;
-    uint8_t character_set;
-    uint32_t status_flags;
-    String auth_plugin_name;
-    String auth_plugin_data;
-
-protected:
-    size_t getPayloadSize() const override;
-
-    void readPayloadImpl(ReadBuffer & payload) override;
-
-    void writePayloadImpl(WriteBuffer & buffer) const override;
-
-public:
-    Handshake();
-
-    Handshake(uint32_t capability_flags_, uint32_t connection_id_, String server_version_, String auth_plugin_name_, String auth_plugin_data_);
-};
-
-class HandshakeResponse : public IMySQLWritePacket, public IMySQLReadPacket
-{
-public:
-    uint32_t capability_flags;
-    uint32_t max_packet_size;
-    uint8_t character_set;
-    String username;
-    String database;
-    String auth_response;
-    String auth_plugin_name;
-
-protected:
-    size_t getPayloadSize() const override;
-
-    void readPayloadImpl(ReadBuffer & payload) override;
-
-    void writePayloadImpl(WriteBuffer & buffer) const override;
-
-public:
-    HandshakeResponse();
-
-    HandshakeResponse(
-        UInt32 capability_flags_, UInt32 max_packet_size_, UInt8 character_set_,
-        const String & username_, const String & database_, const String & auth_response_, const String & auth_plugin_name_);
-};
-
-class AuthSwitchRequest : public IMySQLWritePacket
-{
-public:
-    String plugin_name;
-    String auth_plugin_data;
-
-protected:
-    size_t getPayloadSize() const override;
-
-    void writePayloadImpl(WriteBuffer & buffer) const override;
-
-public:
-    AuthSwitchRequest(String plugin_name_, String auth_plugin_data_);
-};
-
-class AuthSwitchResponse : public LimitedReadPacket
-{
-public:
-    String value;
-
-protected:
-    void readPayloadImpl(ReadBuffer & payload) override;
-};
-
-class AuthMoreData : public IMySQLWritePacket
-{
-public:
-    String data;
-
-protected:
-    size_t getPayloadSize() const override;
-
-    void writePayloadImpl(WriteBuffer & buffer) const override;
-
-public:
-    explicit AuthMoreData(String data_): data(std::move(data_)) {}
 };
 
 class OKPacket : public IMySQLWritePacket, public IMySQLReadPacket
@@ -156,19 +70,6 @@ public:
     EOFPacket(int warnings_, int status_flags_);
 };
 
-class AuthSwitchPacket : public IMySQLReadPacket
-{
-public:
-    String plugin_name;
-
-    AuthSwitchPacket() = default;
-
-protected:
-    UInt8 header = 0x00;
-
-    void readPayloadImpl(ReadBuffer & payload) override;
-};
-
 class ERRPacket : public IMySQLWritePacket, public IMySQLReadPacket
 {
 public:
@@ -188,6 +89,19 @@ public:
     ERRPacket();
 
     ERRPacket(int error_code_, String sql_state_, String error_message_);
+};
+
+class AuthSwitchPacket : public IMySQLReadPacket
+{
+public:
+    String plugin_name;
+
+    AuthSwitchPacket() = default;
+
+protected:
+    UInt8 header = 0x00;
+
+    void readPayloadImpl(ReadBuffer & payload) override;
 };
 
 enum ResponsePacketType
