+++ src/Core/MySQLProtocol.h
@@ -23,6 +23,9 @@
 #include <Poco/RandomStream.h>
 #include <Poco/SHA1Engine.h>
 #include <Core/MySQL/MySQLPackets.h>
+#include <Core/MySQL/PacketPayloadReadBuffer.h>
+#include <Core/MySQL/PacketPayloadWriteBuffer.h>
+#include <Core/MySQL/PacketEndpoint.h>
 
 #if !defined(ARCADIA_BUILD)
 #    include "config_core.h"
@@ -52,7 +55,7 @@ namespace ErrorCodes
 namespace MySQLProtocol
 {
 
-const size_t MAX_PACKET_LENGTH = (1 << 24) - 1; // 16 mb
+//const size_t MAX_PACKET_LENGTH = (1 << 24) - 1; // 16 mb
 const size_t SCRAMBLE_LENGTH = 20;
 const size_t AUTH_PLUGIN_DATA_PART_1_LENGTH = 8;
 const size_t MYSQL_ERRMSG_SIZE = 512;
@@ -125,174 +128,6 @@ public:
 };
 
 
-/** Reading packets.
- *  Internally, it calls (if no more data) next() method of the underlying ReadBufferFromPocoSocket, and sets the working buffer to the rest part of the current packet payload.
- */
-class PacketPayloadReadBuffer : public ReadBuffer
-{
-private:
-    ReadBuffer & in;
-    uint8_t & sequence_id;
-    const size_t max_packet_size = MAX_PACKET_LENGTH;
-
-    bool has_read_header = false;
-
-    // Size of packet which is being read now.
-    size_t payload_length = 0;
-
-    // Offset in packet payload.
-    size_t offset = 0;
-
-protected:
-    bool nextImpl() override;
-
-public:
-    PacketPayloadReadBuffer(ReadBuffer & in_, uint8_t & sequence_id_);
-};
-
-
-/** Writing packets.
- *  https://dev.mysql.com/doc/internals/en/mysql-packet.html
- */
-class PacketPayloadWriteBuffer : public WriteBuffer
-{
-public:
-    PacketPayloadWriteBuffer(WriteBuffer & out_, size_t payload_length_, uint8_t & sequence_id_)
-        : WriteBuffer(out_.position(), 0), out(out_), sequence_id(sequence_id_), total_left(payload_length_)
-    {
-        startNewPacket();
-        setWorkingBuffer();
-        pos = out.position();
-    }
-
-    bool remainingPayloadSize()
-    {
-        return total_left;
-    }
-
-private:
-    WriteBuffer & out;
-    uint8_t & sequence_id;
-
-    size_t total_left = 0;
-    size_t payload_length = 0;
-    size_t bytes_written = 0;
-    bool eof = false;
-
-    void startNewPacket()
-    {
-        payload_length = std::min(total_left, MAX_PACKET_LENGTH);
-        bytes_written = 0;
-        total_left -= payload_length;
-
-        out.write(reinterpret_cast<char *>(&payload_length), 3);
-        out.write(sequence_id++);
-        bytes += 4;
-    }
-
-    /// Sets working buffer to the rest of current packet payload.
-    void setWorkingBuffer()
-    {
-        out.nextIfAtEnd();
-        working_buffer = WriteBuffer::Buffer(out.position(), out.position() + std::min(payload_length - bytes_written, out.available()));
-
-        if (payload_length - bytes_written == 0)
-        {
-            /// Finished writing packet. Due to an implementation of WriteBuffer, working_buffer cannot be empty. Further write attempts will throw Exception.
-            eof = true;
-            working_buffer.resize(1);
-        }
-    }
-
-protected:
-    void nextImpl() override
-    {
-        const int written = pos - working_buffer.begin();
-        if (eof)
-            throw Exception("Cannot write after end of buffer.", ErrorCodes::CANNOT_WRITE_AFTER_END_OF_BUFFER);
-
-        out.position() += written;
-        bytes_written += written;
-
-        /// Packets of size greater than MAX_PACKET_LENGTH are split into few packets of size MAX_PACKET_LENGTH and las packet of size < MAX_PACKET_LENGTH.
-        if (bytes_written == payload_length && (total_left > 0 || payload_length == MAX_PACKET_LENGTH))
-            startNewPacket();
-
-        setWorkingBuffer();
-    }
-};
-
-/* Writes and reads packets, keeping sequence-id.
- * Throws ProtocolError, if packet with incorrect sequence-id was received.
- */
-class PacketSender
-{
-public:
-    uint8_t & sequence_id;
-    ReadBuffer * in;
-    WriteBuffer * out;
-    size_t max_packet_size = MAX_PACKET_LENGTH;
-
-    /// For reading and writing.
-    PacketSender(ReadBuffer & in_, WriteBuffer & out_, uint8_t & sequence_id_)
-        : sequence_id(sequence_id_)
-        , in(&in_)
-        , out(&out_)
-    {
-    }
-
-    /// For writing.
-    PacketSender(WriteBuffer & out_, uint8_t & sequence_id_)
-        : sequence_id(sequence_id_)
-        , in(nullptr)
-        , out(&out_)
-    {
-    }
-
-    void receivePacket(IMySQLReadPacket & packet)
-    {
-        packet.readPayload(*in, sequence_id);
-    }
-
-    bool tryReceivePacket(IMySQLReadPacket & packet, UInt64 millisecond = 0)
-    {
-        if (millisecond != 0)
-        {
-            ReadBufferFromPocoSocket * socket_in = typeid_cast<ReadBufferFromPocoSocket *>(in);
-
-            if (!socket_in)
-                throw Exception("LOGICAL ERROR: Attempt to pull the duration in a non socket stream", ErrorCodes::LOGICAL_ERROR);
-
-            if (!socket_in->poll(millisecond * 1000))
-                return false;
-        }
-
-        packet.readPayload(*in, sequence_id);
-        return true;
-    }
-
-    template<class T>
-    void sendPacket(const T & packet, bool flush = false)
-    {
-        static_assert(std::is_base_of<IMySQLWritePacket, T>());
-        packet.writePayload(*out, sequence_id);
-        if (flush)
-            out->next();
-    }
-
-    PacketPayloadReadBuffer getPayload()
-    {
-        return PacketPayloadReadBuffer(*in, sequence_id);
-    }
-
-    /// Sets sequence-id to 0. Must be called before each command phase.
-    void resetSequenceId();
-
-    /// Converts packet to text. Is used for debug output.
-    static String packetToText(const String & payload);
-};
-
-
 uint64_t readLengthEncodedNumber(ReadBuffer & ss);
 
 inline void readLengthEncodedString(String & s, ReadBuffer & buffer)
@@ -321,325 +156,6 @@ size_t getLengthEncodedNumberSize(uint64_t x);
 size_t getLengthEncodedStringSize(const String & s);
 
 ColumnDefinitionPacket getColumnDefinition(const String & column_name, const TypeIndex index);
-
-
-namespace ProtocolText
-{
-
-class ResultsetRow : public IMySQLWritePacket
-{
-    const Columns & columns;
-    int row_num;
-    size_t payload_size = 0;
-    std::vector<String> serialized;
-public:
-    ResultsetRow(const DataTypes & data_types, const Columns & columns_, int row_num_)
-        : columns(columns_)
-        , row_num(row_num_)
-    {
-        for (size_t i = 0; i < columns.size(); i++)
-        {
-            if (columns[i]->isNullAt(row_num))
-            {
-                payload_size += 1;
-                serialized.emplace_back("\xfb");
-            }
-            else
-            {
-                WriteBufferFromOwnString ostr;
-                data_types[i]->serializeAsText(*columns[i], row_num, ostr, FormatSettings());
-                payload_size += getLengthEncodedStringSize(ostr.str());
-                serialized.push_back(std::move(ostr.str()));
-            }
-        }
-    }
-protected:
-    size_t getPayloadSize() const override
-    {
-        return payload_size;
-    }
-
-    void writePayloadImpl(WriteBuffer & buffer) const override
-    {
-        for (size_t i = 0; i < columns.size(); i++)
-        {
-            if (columns[i]->isNullAt(row_num))
-                buffer.write(serialized[i].data(), 1);
-            else
-                writeLengthEncodedString(serialized[i], buffer);
-        }
-    }
-};
-
-}
-
-namespace Authentication
-{
-
-class IPlugin
-{
-public:
-    virtual String getName() = 0;
-
-    virtual String getAuthPluginData() = 0;
-
-    virtual void authenticate(const String & user_name, std::optional<String> auth_response, Context & context, std::shared_ptr<PacketSender> packet_sender, bool is_secure_connection,
-                              const Poco::Net::SocketAddress & address) = 0;
-
-    virtual ~IPlugin() = default;
-};
-
-/// https://dev.mysql.com/doc/internals/en/secure-password-authentication.html
-class Native41 : public IPlugin
-{
-public:
-    Native41()
-    {
-        scramble.resize(SCRAMBLE_LENGTH + 1, 0);
-        Poco::RandomInputStream generator;
-
-       /** Generate a random string using ASCII characters but avoid separator character,
-         * produce pseudo random numbers between with about 7 bit worth of entropty between 1-127.
-         * https://github.com/mysql/mysql-server/blob/8.0/mysys/crypt_genhash_impl.cc#L427
-         */
-        for (size_t i = 0; i < SCRAMBLE_LENGTH; ++i)
-        {
-            generator >> scramble[i];
-            scramble[i] &= 0x7f;
-            if (scramble[i] == '\0' || scramble[i] == '$')
-                scramble[i] = scramble[i] + 1;
-        }
-    }
-
-    Native41(const String & password, const String & auth_plugin_data)
-    {
-        /// https://dev.mysql.com/doc/internals/en/secure-password-authentication.html
-        /// SHA1( password ) XOR SHA1( "20-bytes random data from server" <concat> SHA1( SHA1( password ) ) )
-        Poco::SHA1Engine engine1;
-        engine1.update(password.data());
-        const Poco::SHA1Engine::Digest & password_sha1 = engine1.digest();
-
-        Poco::SHA1Engine engine2;
-        engine2.update(password_sha1.data(), password_sha1.size());
-        const Poco::SHA1Engine::Digest & password_double_sha1 = engine2.digest();
-
-        Poco::SHA1Engine engine3;
-        engine3.update(auth_plugin_data.data(), auth_plugin_data.size());
-        engine3.update(password_double_sha1.data(), password_double_sha1.size());
-        const Poco::SHA1Engine::Digest & digest = engine3.digest();
-
-        scramble.resize(SCRAMBLE_LENGTH);
-        for (size_t i = 0; i < SCRAMBLE_LENGTH; i++)
-        {
-            scramble[i] = static_cast<unsigned char>(password_sha1[i] ^ digest[i]);
-        }
-    }
-
-    String getName() override
-    {
-        return "mysql_native_password";
-    }
-
-    String getAuthPluginData() override
-    {
-        return scramble;
-    }
-
-    void authenticate(
-        const String & user_name,
-        std::optional<String> auth_response,
-        Context & context,
-        std::shared_ptr<PacketSender> packet_sender,
-        bool /* is_secure_connection */,
-        const Poco::Net::SocketAddress & address) override
-    {
-        if (!auth_response)
-        {
-            packet_sender->sendPacket(AuthSwitchRequest(getName(), scramble), true);
-            AuthSwitchResponse response;
-            packet_sender->receivePacket(response);
-            auth_response = response.value;
-        }
-
-        if (auth_response->empty())
-        {
-            context.setUser(user_name, "", address);
-            return;
-        }
-
-        if (auth_response->size() != Poco::SHA1Engine::DIGEST_SIZE)
-            throw Exception("Wrong size of auth response. Expected: " + std::to_string(Poco::SHA1Engine::DIGEST_SIZE) + " bytes, received: " + std::to_string(auth_response->size()) + " bytes.",
-                            ErrorCodes::UNKNOWN_EXCEPTION);
-
-        auto user = context.getAccessControlManager().read<User>(user_name);
-
-        Poco::SHA1Engine::Digest double_sha1_value = user->authentication.getPasswordDoubleSHA1();
-        assert(double_sha1_value.size() == Poco::SHA1Engine::DIGEST_SIZE);
-
-        Poco::SHA1Engine engine;
-        engine.update(scramble.data(), SCRAMBLE_LENGTH);
-        engine.update(double_sha1_value.data(), double_sha1_value.size());
-
-        String password_sha1(Poco::SHA1Engine::DIGEST_SIZE, 0x0);
-        const Poco::SHA1Engine::Digest & digest = engine.digest();
-        for (size_t i = 0; i < password_sha1.size(); i++)
-        {
-            password_sha1[i] = digest[i] ^ static_cast<unsigned char>((*auth_response)[i]);
-        }
-        context.setUser(user_name, password_sha1, address);
-    }
-private:
-    String scramble;
-};
-
-#if USE_SSL
-/// Caching SHA2 plugin is not used because it would be possible to authenticate knowing hash from users.xml.
-/// https://dev.mysql.com/doc/internals/en/sha256.html
-class Sha256Password : public IPlugin
-{
-public:
-    Sha256Password(RSA & public_key_, RSA & private_key_, Poco::Logger * log_)
-        : public_key(public_key_)
-        , private_key(private_key_)
-        , log(log_)
-    {
-        /** Native authentication sent 20 bytes + '\0' character = 21 bytes.
-         *  This plugin must do the same to stay consistent with historical behavior if it is set to operate as a default plugin. [1]
-         *  https://github.com/mysql/mysql-server/blob/8.0/sql/auth/sql_authentication.cc#L3994
-         */
-        scramble.resize(SCRAMBLE_LENGTH + 1, 0);
-        Poco::RandomInputStream generator;
-
-        for (size_t i = 0; i < SCRAMBLE_LENGTH; ++i)
-        {
-            generator >> scramble[i];
-            scramble[i] &= 0x7f;
-            if (scramble[i] == '\0' || scramble[i] == '$')
-                scramble[i] = scramble[i] + 1;
-        }
-    }
-
-    String getName() override
-    {
-        return "sha256_password";
-    }
-
-    String getAuthPluginData() override
-    {
-        return scramble;
-    }
-
-    void authenticate(
-        const String & user_name,
-        std::optional<String> auth_response,
-        Context & context,
-        std::shared_ptr<PacketSender> packet_sender,
-        bool is_secure_connection,
-        const Poco::Net::SocketAddress & address) override
-    {
-        if (!auth_response)
-        {
-            packet_sender->sendPacket(AuthSwitchRequest(getName(), scramble), true);
-
-            if (packet_sender->in->eof())
-                throw Exception("Client doesn't support authentication method " + getName() + " used by ClickHouse. Specifying user password using 'password_double_sha1_hex' may fix the problem.",
-                    ErrorCodes::MYSQL_CLIENT_INSUFFICIENT_CAPABILITIES);
-
-            AuthSwitchResponse response;
-            packet_sender->receivePacket(response);
-            auth_response = response.value;
-            LOG_TRACE(log, "Authentication method mismatch.");
-        }
-        else
-        {
-            LOG_TRACE(log, "Authentication method match.");
-        }
-
-        bool sent_public_key = false;
-        if (auth_response == "\1")
-        {
-            LOG_TRACE(log, "Client requests public key.");
-            BIO * mem = BIO_new(BIO_s_mem());
-            SCOPE_EXIT(BIO_free(mem));
-            if (PEM_write_bio_RSA_PUBKEY(mem, &public_key) != 1)
-            {
-                throw Exception("Failed to write public key to memory. Error: " + getOpenSSLErrors(), ErrorCodes::OPENSSL_ERROR);
-            }
-            char * pem_buf = nullptr;
-#    pragma GCC diagnostic push
-#    pragma GCC diagnostic ignored "-Wold-style-cast"
-            long pem_size = BIO_get_mem_data(mem, &pem_buf);
-#    pragma GCC diagnostic pop
-            String pem(pem_buf, pem_size);
-
-            LOG_TRACE(log, "Key: {}", pem);
-
-            AuthMoreData data(pem);
-            packet_sender->sendPacket(data, true);
-            sent_public_key = true;
-
-            AuthSwitchResponse response;
-            packet_sender->receivePacket(response);
-            auth_response = response.value;
-        }
-        else
-        {
-            LOG_TRACE(log, "Client didn't request public key.");
-        }
-
-        String password;
-
-        /** Decrypt password, if it's not empty.
-         *  The original intention was that the password is a string[NUL] but this never got enforced properly so now we have to accept that
-         *  an empty packet is a blank password, thus the check for auth_response.empty() has to be made too.
-         *  https://github.com/mysql/mysql-server/blob/8.0/sql/auth/sql_authentication.cc#L4017
-         */
-        if (!is_secure_connection && !auth_response->empty() && auth_response != String("\0", 1))
-        {
-            LOG_TRACE(log, "Received nonempty password.");
-            auto ciphertext = reinterpret_cast<unsigned char *>(auth_response->data());
-
-            unsigned char plaintext[RSA_size(&private_key)];
-            int plaintext_size = RSA_private_decrypt(auth_response->size(), ciphertext, plaintext, &private_key, RSA_PKCS1_OAEP_PADDING);
-            if (plaintext_size == -1)
-            {
-                if (!sent_public_key)
-                    LOG_WARNING(log, "Client could have encrypted password with different public key since it didn't request it from server.");
-                throw Exception("Failed to decrypt auth data. Error: " + getOpenSSLErrors(), ErrorCodes::OPENSSL_ERROR);
-            }
-
-            password.resize(plaintext_size);
-            for (int i = 0; i < plaintext_size; i++)
-            {
-                password[i] = plaintext[i] ^ static_cast<unsigned char>(scramble[i % scramble.size()]);
-            }
-        }
-        else if (is_secure_connection)
-        {
-            password = *auth_response;
-        }
-        else
-        {
-            LOG_TRACE(log, "Received empty password");
-        }
-
-        if (!password.empty() && password.back() == 0)
-        {
-            password.pop_back();
-        }
-
-        context.setUser(user_name, password, address);
-    }
-
-private:
-    RSA & public_key;
-    RSA & private_key;
-    Poco::Logger * log;
-    String scramble;
-};
-#endif
-
-}
 
 namespace Replication
 {
