+++ src/Processors/Formats/Impl/MySQLOutputFormat.h
@@ -26,7 +26,7 @@ public:
     void setContext(const Context & context_)
     {
         context = &context_;
-        packet_sender = std::make_unique<MySQLProtocol::PacketSender>(out, const_cast<uint8_t &>(context_.mysql.sequence_id)); /// TODO: fix it
+        packet_sender = std::make_unique<MySQLProtocol::PacketEndpoint>(out, const_cast<uint8_t &>(context_.mysql.sequence_id)); /// TODO: fix it
         packet_sender->max_packet_size = context_.mysql.max_packet_size;
     }
 
@@ -42,7 +42,7 @@ private:
     bool initialized = false;
 
     const Context * context = nullptr;
-    std::unique_ptr<MySQLProtocol::PacketSender> packet_sender;
+    std::unique_ptr<MySQLProtocol::PacketEndpoint> packet_sender;
     FormatSettings format_settings;
     DataTypes data_types;
 };
