+++ src/Server/MySQLHandler.h
@@ -33,7 +33,7 @@ private:
     CurrentMetrics::Increment metric_increment{CurrentMetrics::MySQLConnection};
 
     /// Enables SSL, if client requested.
-    void finishHandshake(MySQLProtocol::HandshakeResponse &);
+    void finishHandshake(MySQLProtocol::ConnectionPhase::HandshakeResponse &);
 
     void comQuery(ReadBuffer & payload);
 
@@ -46,7 +46,7 @@ private:
     void authenticate(const String & user_name, const String & auth_plugin_name, const String & auth_response);
 
     virtual void authPluginSSL();
-    virtual void finishHandshakeSSL(size_t packet_size, char * buf, size_t pos, std::function<void(size_t)> read_bytes, MySQLProtocol::HandshakeResponse & packet);
+    virtual void finishHandshakeSSL(size_t packet_size, char * buf, size_t pos, std::function<void(size_t)> read_bytes, MySQLProtocol::ConnectionPhase::HandshakeResponse & packet);
 
     IServer & server;
 
@@ -55,7 +55,7 @@ protected:
 
     Context connection_context;
 
-    std::shared_ptr<MySQLProtocol::PacketSender> packet_sender;
+    std::shared_ptr<MySQLProtocol::PacketEndpoint> packet_sender;
 
 private:
     size_t connection_id = 0;
@@ -85,7 +85,10 @@ public:
 
 private:
     void authPluginSSL() override;
-    void finishHandshakeSSL(size_t packet_size, char * buf, size_t pos, std::function<void(size_t)> read_bytes, MySQLProtocol::HandshakeResponse & packet) override;
+
+    void finishHandshakeSSL(
+        size_t packet_size, char * buf, size_t pos,
+        std::function<void(size_t)> read_bytes, MySQLProtocol::ConnectionPhase::HandshakeResponse & packet) override;
 
     RSA & public_key;
     RSA & private_key;
